Closes VP-204

+++ src/main/java/no/ntnu/bachelor/voicepick/features/pluck/dtos/CargoCarrierDto.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @AllArgsConstructor
 public class CargoCarrierDto {
   private String name;
-  private Long identifier;
+  private int identifier;
   private String phoneticIdentifier;
 }
